refactor(mcp): split personal-state registration
Problem: the mutation MCP registrar combined independent durable personal-state records with cross-cutting write tools, making its execution boundary difficult to navigate and change safely.\n\nWhat changed: move annotations, saved views, recall packs, reader workspaces, metadata, and learning corrections behind one composed personal-state registrar. Keep the original parent as the sole mutation registration route and extract only its shared session-resolution and paging primitives. Contract discovery now follows that composition and verifies removing the composition call removes every extracted production tool.\n\nCompatibility: MCP tool names, role gating, handlers, payloads, error envelopes, and operation code references are preserved.\n\nRef polylogue-1r9c.
